// src/types/PlayerGameweekData.ts
interface PlayerGameweekData {
    fpl_minutes: number;
    fpl_round: number;
    fpl_total_points: number;
    fpl_goals_scored: number;
    fpl_assists: number;
    fpl_bps: number;
    fpl_clean_sheet: number;
    fpl_defensive_contribution: number;
    fpl_expected_goals: number;
    fpl_expected_assists: number;
    fpl_xgi: number;
    sm_shots_on_target: number;
    sm_big_chances_created: number;
    sm_key_passes: number;
    calc_xgap: number;
}

export default PlayerGameweekData;